move migrations to their own module
